foi feita a modularizacion porem tem que ser testada no html
